Fix Playback management regressions and bugs

+++ app/src/main/java/se/splushii/dancingbunnies/storage/db/PlaybackControllerEntry.java
@@ -46,7 +46,7 @@ public class PlaybackControllerEntry {
     int queueID;
     @NonNull
     @ColumnInfo(name = COLUMN_POS)
-    long pos;
+    int pos;
     @NonNull
     @ColumnInfo(name = COLUMN_API)
     public String api;
